requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.30.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
